Move List tests into the Module tests.
